Refactor todo item type to new types file
