+++ src/App.tsx
@@ -11,16 +11,6 @@ export type User = {
   token: string;
 };
 
-export type TodoItem = {
-  id: string;
-  title: string;
-  description?: string;
-  completed: boolean;
-  tags?: string[];
-  reminderDate?: Date;
-  reminderTime?: Date;
-};
-
 function App() {
   return (
     <Router>
