+++ src/types/todo.ts
@@ -1,0 +1,9 @@
+export type TodoItem = {
+  id: string;
+  title: string;
+  description?: string;
+  completed: boolean;
+  tags?: string[];
+  reminderDate?: Date;
+  reminderTime?: Date;
+};
